requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==4.2.11
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
filelock==3.14.0
identify==2.5.36
iniconfig==2.0.0
isort==5.13.2
nodeenv==1.8.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
virtualenv==20.26.1
